ListInstrumentGroupMeasurements; refactor; docs
