+++ tests/instrumentation-regression.postman_collection.json
@@ -1,6 +1,6 @@
 {
 	"info": {
-		"_postman_id": "b490dde5-1e24-4191-a786-ba0dd150c134",
+		"_postman_id": "51d89327-5896-4fe4-82b8-ccd27573c248",
 		"name": "instrumentation-regression",
 		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
 	},
@@ -1986,6 +1986,61 @@
 									"key": "instrument_group_id",
 									"value": "d0916e8a-39a6-4f2f-bd31-879881f8b40c",
 									"description": "Instrument Group uuid"
+								}
+							]
+						}
+					},
+					"response": []
+				},
+				{
+					"name": "ListInstrumentGroupMeasurements",
+					"event": [
+						{
+							"listen": "test",
+							"script": {
+								"exec": [
+									"// status code is 200",
+									"pm.test(\"Status code is 200\", function () {",
+									"    pm.response.to.have.status(200);",
+									"});",
+									""
+								],
+								"type": "text/javascript"
+							}
+						}
+					],
+					"request": {
+						"method": "GET",
+						"header": [],
+						"url": {
+							"raw": "{{base_url}}/instrument_groups/:instrument_group_id/timeseries_measurements?after=1900-01-01T00:00:00.00Z&before=2025-12-31T00:00:00.00Z&interval=",
+							"host": [
+								"{{base_url}}"
+							],
+							"path": [
+								"instrument_groups",
+								":instrument_group_id",
+								"timeseries_measurements"
+							],
+							"query": [
+								{
+									"key": "after",
+									"value": "1900-01-01T00:00:00.00Z"
+								},
+								{
+									"key": "before",
+									"value": "2025-12-31T00:00:00.00Z"
+								},
+								{
+									"key": "interval",
+									"value": ""
+								}
+							],
+							"variable": [
+								{
+									"key": "instrument_group_id",
+									"value": "d0916e8a-39a6-4f2f-bd31-879881f8b40c",
+									"description": "instrument group uuid"
 								}
 							]
 						}
